refactor(dashboard): organize analytics feature
