+++ dashboard/src/features/analytics/user.ts
@@ -1,0 +1,1 @@
+export { UserAnalyticsDrawer } from "./user-analytics-drawer"
